Minor fixes and refactoring based on unit tests
