+++ packages/google_maps_routes_api/lib/src/types/route_matrix_request.dart
@@ -1,8 +1,9 @@
 // Copyright 2013 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
+
 import 'enums.dart';
-import 'routes_request.dart';
+import 'route_modifiers.dart';
 import 'waypoint.dart';
 
 /// Takes in a list of [origins] and [destinations] and returns route
